Some refactoring increasing readability
